Woot! 1 second faster.
This change reduces the amount that's loaded up front when the processor is spawned. Because we were loading ALL the processors (and require blocks + all the require'd deps) it was adding load time.

By requiring the requested processor on demand, our render time for an example bin goes from 1.2-1.5 to 0.3 seconds.